Add files via upload
The objective of this project is to identify recent research trends in operations research using text mining. To do this, a collection of abstracts of research papers presented at  three different Operations Research conferences around the world(INFORMS, EURO and GOR) in 2022  was mined. The python notebooks contain the LDA and BERTopic model implementations with Python programming.